rename: enum 파일 package 이동
