+++ src/main/java/com/capstone/rentit/item/dto/ItemDtoFactory.java
@@ -1,16 +1,6 @@
 package com.capstone.rentit.item.dto;
 
-import com.capstone.rentit.common.ItemStatusConverter;
-import com.capstone.rentit.common.MemberRoleEnum;
 import com.capstone.rentit.item.domain.Item;
-import com.capstone.rentit.member.domain.Company;
-import com.capstone.rentit.member.domain.Member;
-import com.capstone.rentit.member.domain.Student;
-import com.capstone.rentit.member.domain.StudentCouncilMember;
-import com.capstone.rentit.member.dto.CompanyDto;
-import com.capstone.rentit.member.dto.MemberDto;
-import com.capstone.rentit.member.dto.StudentCouncilMemberDto;
-import com.capstone.rentit.member.dto.StudentDto;
 
 import java.time.LocalDateTime;
