+++ src/main/java/com/capstone/rentit/item/status/ItemStatusConverter.java
@@ -1,4 +1,4 @@
-package com.capstone.rentit.common;
+package com.capstone.rentit.item.status;
 
 public class ItemStatusConverter {
