+++ src/main/java/com/capstone/rentit/item/status/ItemStatusEnum.java
@@ -1,4 +1,4 @@
-package com.capstone.rentit.common;
+package com.capstone.rentit.item.status;
 
 public enum ItemStatusEnum {
     OUT, AVAILABLE;
